% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonLinearDotPlot.R
\name{dotscaling.linear}
\alias{dotscaling.linear}
\title{Predefined function to use as \code{dotscaling} in the \code{nonLinearDotPlot} function.}
\usage{
dotscaling.linear()
}
\value{
Function that always returns 1.
}
\description{
dotscaling(c) = 1 is a linear function
}
\references{
N. Rodrigues and D. Weiskopf, "Nonlinear Dot Plots",
IEEE Transactions on Visualization and Computer Graphics, vol. 24, no. 1, pp. 616-625, 2018.
Available: \doi{10.1109/TVCG.2017.2744018}
}
